share router between adapter tests
